requirements.txt
requests
Flask
Flask-CORS
Flask-SQLAlchemy
pymysql
pyyaml